+++ template-parts/primary-nav/toggler.php
@@ -13,5 +13,7 @@
 
 namespace Secretum;
 
-?>
+if ( false !== secretum_mod( 'primary_nav_toggler_status' ) ) { ?>
 <button class="navbar-toggler<?php secretum_nav_toggler_wrapper( 'primary_nav' ); ?>" type="button" data-toggle="collapse" data-target="#navbarNavDropdown" aria-controls="navbarNavDropdown" aria-expanded="false" aria-label="Toggle navigation"><span class="navbar-toggler-icon<?php secretum_nav_toggler_icon( 'primary_nav' ); ?>"></span></button>
+	<?php
+}